refactor: move kafka utils file
