+++ services/collab-service/src/kafka/utils.ts
